fixed Sarah's merged conflict errors/updated CSS
