+++ src/App.css
@@ -1,43 +1,46 @@
 html{line-height:1.15;-ms-text-size-adjust:100%;-webkit-text-size-adjust:100%}body{margin:0}article,aside,footer,header,nav,section{display:block}h1{font-size:2em;margin:.67em 0}figcaption,figure,main{display:block}figure{margin:1em 40px}hr{box-sizing:content-box;height:0;overflow:visible}pre{font-family:monospace,monospace;font-size:1em}a{background-color:transparent;-webkit-text-decoration-skip:objects}abbr[title]{border-bottom:none;text-decoration:underline;text-decoration:underline dotted}b,strong{font-weight:inherit}b,strong{font-weight:bolder}code,kbd,samp{font-family:monospace,monospace;font-size:1em}dfn{font-style:italic}mark{background-color:#ff0;color:#000}small{font-size:80%}sub,sup{font-size:75%;line-height:0;position:relative;vertical-align:baseline}sub{bottom:-.25em}sup{top:-.5em}audio,video{display:inline-block}audio:not([controls]){display:none;height:0}img{border-style:none}svg:not(:root){overflow:hidden}button,input,optgroup,select,textarea{font-family:sans-serif;font-size:100%;line-height:1.15;margin:0}button,input{overflow:visible}button,select{text-transform:none}button,html [type=button],[type=reset],[type=submit]{-webkit-appearance:button}button::-moz-focus-inner,[type=button]::-moz-focus-inner,[type=reset]::-moz-focus-inner,[type=submit]::-moz-focus-inner{border-style:none;padding:0}button:-moz-focusring,[type=button]:-moz-focusring,[type=reset]:-moz-focusring,[type=submit]:-moz-focusring{outline:1px dotted ButtonText}fieldset{padding:.35em .75em .625em}legend{box-sizing:border-box;color:inherit;display:table;max-width:100%;padding:0;white-space:normal}progress{display:inline-block;vertical-align:baseline}textarea{overflow:auto}[type=checkbox],[type=radio]{box-sizing:border-box;padding:0}[type=number]::-webkit-inner-spin-button,[type=number]::-webkit-outer-spin-button{height:auto}[type=search]{-webkit-appearance:textfield;outline-offset:-2px}[type=search]::-webkit-search-cancel-button,[type=search]::-webkit-search-decoration{-webkit-appearance:none}::-webkit-file-upload-button{-webkit-appearance:button;font:inherit}details,menu{display:block}summary{display:list-item}canvas{display:inline-block}template{display:none}[hidden]{display:none}
-
 .clearfix:after {visibility: hidden; display: block; font-size: 0; content: ''; clear: both; height: 0; }
-
 html { -moz-box-sizing: border-box; -webkit-box-sizing: border-box; box-sizing: border-box;}
-
 *, *:before, *:after {box-sizing: inherit}
-
 .sr-only { position: absolute; width: 1px; height: 1px; margin: -1px; border: 0; padding: 0; white-space: nowrap; clip-path: inset(100%); clip: rect(0 0 0 0); overflow: hidden;}
-
 :root {
-  --font: ‘Montserrat’, sans-serif; 
+  --font: ‘Montserrat’, sans-serif;
   --dark: #577F90;
   --light: #86C2AE;
   --accent: #F0917A;
   --textColor: black;
 }
+/* 
+* {
+  margin: 0;
+  padding: 0;
+} */
 
-/* Font Family */
-
+/* 
+section,
+main,
+div,
+h1,
+h2,
+h3,
+p,
+input,
+labels,
+textarea */
 /* Header */
-
 html {
   font-size: 62.5%;
   font-family: ‘Montserrat’, sans-serif;
 }
 
-.App {
-  display: flex;
-  height: 100vh;
-  width: 100vw;
-  min-width: 100vh;
-}
-
 header {
-  display: flex;
-  background-image: url(./assets/microphone.jpg);
-  width: 60%;
-  height: 100%;
-  background-size: 100%;
+  background-image: url(./assets/headerBanner.jpg);
+  background-size: contain;
+  background-repeat: no-repeat;
+  background-position: center;
+  background-position-y: top;
+
+  height: 90vh; 
 }
 
 h1 {
@@ -45,41 +48,137 @@ h1 {
   border-bottom: green;
   text-align: center;
 }
-
 /* Logo goes here */
-
-body {
+/* BODY */
+main {
+  /* display: flex; */
   background-color: white;
-  width: 40%;
-  height: 100%;
+  width: 100%;
+  height: 100vh;
 }
-
 .wrapper {
-  max-width: 1080px;
+  max-width: 1280px;
   margin: 0 auto;
-  width: 30%;
+  width: 90%;
 }
 
+.mainContainer {
+  display: flex;
+  justify-content: space-between;
+}
+
+.left {
+  width: 38%;
+}
+
+h3 {
+  font-size: 2rem;
+  /* text-align: center; */
+}
+.logoContainer {
+  height: 50px;
+  width: 50px;
+  background-image: url(./assets/myLyricsButton.jpg);
+  background-size: 100% 100%;
+  position: relative;
+  top: 55px;
+  z-index: 1;
+}
+/* Having trouble applying images unless they're backgrounds */
+/* Left */
+.findLyrics,
+.myLyrics {
+  display: flex;
+  flex-direction: column;
+  /* height: 50%; */
+  border: 3px solid #577F90;
+  /* margin: 10px 0; */
+  border-radius: 10px;;
+  padding: 0 20px;
+}
+.findLyrics {
+  height: 35%;
+}
+.myLyrics {
+  height: 60%;
+}
+.findLyrics h3 {
+  text-align: center;
+}
+.myLyrics h3 {
+  border-bottom: #F0917A solid 4px;
+  position: relative;
+  padding-left: 60px;
+  width: 100%;
+}
+.savedSongs {
+  display: flex;
+  padding: 5px;
+}
+.savedSongs:nth-of-type(2n+1){
+  background: #86C2AE;
+}
 form {
   display: flex;
   flex-direction: column;
+  width: 100%;
 }
-
+.song,
+.artist {
+  display: flex;
+  justify-content: space-between;
+  margin: 1px 0;
+}
+.song input,
+.artist input {
+  width: 80%;
+}
+label {
+  font-size: 1.5rem;
+}
 input {
   border: 1px solid #577F90;
+  font-size: 1.5rem;
+  width: 10%;
 }
-
+.buttonContainer {
+  display: flex;
+  justify-content: flex-end;
+}
 button {
   margin: 10px 0;
+  background-color: #F0917A;
+  color: white;
+  width: 50%;
+  font-size: 1.5rem;
+  border: none;
+  box-shadow: 2px 2px grey;
+  padding: 3px;
+  margin-right: 2px;
 }
-
 .info {
   height: 30%;
 }
 
+
+/* Lyrics Section */
+
+.right {
+  width: 58%;
+  height: 100%;
+}
+
 .lyrics {
   border: solid 5px #577F90;
+  border-radius: 10px;
   background: #86C2AE;
-  padding: 10px;
-  height: 60%;
+  padding: 15px;
+  height: 90%;
+}
+
+/* Lyrics Textarea */
+.lyricsContainer {
+  resize: none;
+
+  font-size: 1.5rem;
 }
