+++ src/Spinner.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Spinner() {
+    return (
+        <div className="fa-3x">
+            <i className="fad fa-compact-disc fa-spin"  />
+        </div>
+        // <span>pownvpownpoevjpeowjpoej</span>
+    );
+}
+export default Spinner;
